# shared/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "result": {
            "$ref": "#/components/schemas/ResultOut"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "url",
          "status",
          "progress",
          "current_step",
          "error",
          "created_at",
          "updated_at",
          "result"
        ],
        "title": "AnalysisOut",
        "type": "object"
      },
      "CheckerLeadRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "submission_id",
          "email"
        ],
        "title": "CheckerLeadRequest",
        "type": "object"
      },
      "CheckerSubmitRequest": {
        "description": "A public checker submit. brand+category must be non-empty after trim.",
        "properties": {
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "lang": {
            "default": "en",
            "title": "Lang",
            "type": "string"
          }
        },
        "required": [
          "brand",
          "category"
        ],
        "title": "CheckerSubmitRequest",
        "type": "object"
      },
      "CheckerSubmitResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "submission_id": {
            "format": "uuid",
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "submission_id"
        ],
        "title": "CheckerSubmitResponse",
        "type": "object"
      },
      "CompetitorMention": {
        "description": "A competitor brand that showed up in the answers and how many answers\nnamed it (P5.3) \u2014 a proper-noun co-mention over the raw text, not the KYC\ncompetitor list.",
        "properties": {
          "mentions": {
            "title": "Mentions",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "mentions"
        ],
        "title": "CompetitorMention",
        "type": "object"
      },
      "CreateAnalysisRequest": {
        "properties": {
          "url": {
            "format": "uri",
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreateAnalysisRequest",
        "type": "object"
      },
      "CreateAnalysisResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "CreateAnalysisResponse",
        "type": "object"
      },
      "EnginePresence": {
        "description": "One engine's presence in a checker run: ``mentioned`` of ``total`` answers\nnamed the searched brand (P5.3). Read-time aggregate of the ``footprint``\nbooleans; the per-engine totals sum to ``total_responses``.",
        "properties": {
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "mentioned": {
            "title": "Mentioned",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "engine",
          "mentioned",
          "total"
        ],
        "title": "EnginePresence",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PromptOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "category"
        ],
        "title": "PromptOut",
        "type": "object"
      },
      "ResponseOut": {
        "properties": {
          "cost_usd": {
            "title": "Cost Usd",
            "type": "number"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "footprint": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Footprint"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "matched_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Snippet"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_id": {
            "format": "uuid",
            "title": "Prompt Id",
            "type": "string"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "prompt_id",
          "engine",
          "model",
          "raw_text",
          "footprint",
          "matched_snippet",
          "cost_usd"
        ],
        "title": "ResponseOut",
        "type": "object"
      },
      "ResultOut": {
        "properties": {
          "competitors_appeared": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CompetitorMention"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competitors Appeared"
          },
          "engine_presence": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EnginePresence"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Presence"
          },
          "footprint_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Footprint Count"
          },
          "geo_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Score"
          },
          "kyc": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kyc"
          },
          "prompts": {
            "items": {
              "$ref": "#/components/schemas/PromptOut"
            },
            "title": "Prompts",
            "type": "array"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/ResponseOut"
            },
            "title": "Responses",
            "type": "array"
          },
          "total_responses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Responses"
          }
        },
        "required": [
          "kyc",
          "prompts",
          "responses",
          "geo_score",
          "footprint_count",
          "total_responses",
          "engine_presence",
          "competitors_appeared"
        ],
        "title": "ResultOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WaitlistRequest": {
        "description": "A public waitlist signup. The email is validated + normalized server-side\n(trim + lowercase) so a malformed address is a 422 before any row is written,\nand the stored value matches the unique lowercased column.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "WaitlistRequest",
        "type": "object"
      },
      "WaitlistResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "WaitlistResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Yanki API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "post": {
        "operationId": "submit_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "operationId": "read_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Analysis",
        "tags": [
          "analyses"
        ]
      }
    },
    "/api/v1/checker": {
      "post": {
        "operationId": "submit_checker_api_v1_checker_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckerSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckerSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Checker",
        "tags": [
          "analyses"
        ]
      }
    },
    "/api/v1/checker/leads": {
      "post": {
        "operationId": "submit_checker_lead_api_v1_checker_leads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckerLeadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Submit Checker Lead Api V1 Checker Leads Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Checker Lead",
        "tags": [
          "analyses"
        ]
      }
    },
    "/api/v1/waitlist": {
      "post": {
        "operationId": "join_waitlist_api_v1_waitlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Join Waitlist",
        "tags": [
          "analyses"
        ]
      }
    },
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    }
  }
}